fixing gitignore and additional edit for update crud

+++ java/src/main/java/com/example/karyawan/repository/KaryawanRepository.java
@@ -1,6 +1,5 @@
 package com.example.karyawan.repository;
-
 import com.example.karyawan.model.Karyawan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KaryawanRepository extends JpaRepository<Karyawan, Long> {}
+public interface KaryawanRepository extends JpaRepository<Karyawan, Integer> {}
